+++ internal/notice/service.go
@@ -5,7 +5,7 @@ import "context"
 type Service interface {
 	CreateNotice(ctx context.Context, n *Notice) (*Notice, error)
 	GetNotice(ctx context.Context, id string) (*Notice, error)
-	GetNotices(ctx context.Context, page, limit int) ([]*Notice, error)
+	GetNotices(ctx context.Context, page, limit int64) ([]*Notice, error)
 	UpdateNotice(ctx context.Context, params *NoticeUpdateParams) error
 	DeleteNotice(ctx context.Context, id string) error
 }
